# app/controllers/api/v1/hangars_controller.rb
# frozen_string_literal: true

module Api
  module V1
    class HangarsController < ::Api::BaseController
      include HangarFiltersConcern

      before_action :authenticate_user!, only: []
      before_action -> { doorkeeper_authorize! "hangar", "hangar:read" },
        unless: :user_signed_in?,
        only: %i[show export items]
      before_action -> { doorkeeper_authorize! "hangar", "hangar:write" },
        unless: :user_signed_in?,
        except: %i[show export items]

      after_action -> { pagination_header(:vehicles) }, only: %i[show]

      def show
        authorize! with: ::HangarPolicy

        scope = authorized_scope(Vehicle.all).visible.purchased

        if price_range.present?
          vehicle_query_params["sorts"] = "model_price asc"
          scope = scope.includes(:model).where(models: {price: price_range})
        end

        if pledge_price_range.present?
          vehicle_query_params["sorts"] = "model_last_pledge_price asc"
          scope = scope.includes(:model).where(models: {last_pledge_price: pledge_price_range})
        end

        scope = loaner_included?(scope)
        scope = scope.where(models: {cargo: 0.1..}) if vehicle_query_params.delete("with_cargo")
        scope = will_it_fit?(scope) if vehicle_query_params["will_it_fit"].present?

        vehicle_query_params["sorts"] = sorting_params(Vehicle)

        @q = scope.ransack(vehicle_query_params)

        result = @q.result(distinct: true)
          .includes(:vehicle_upgrades, :model_paint, :model_upgrades, model: [:manufacturer])
          .joins(model: [:manufacturer])

        @vehicles = result_with_pagination(result, per_page(Vehicle))
      end

      def destroy
        authorize! with: ::HangarPolicy

        Vehicle.transaction do
          # rubocop:disable Rails/SkipsModelValidations
          authorized_scope(Vehicle.all).purchased.update_all(notify: false)
          # rubocop:enable Rails/SkipsModelValidations

          vehicle_ids = authorized_scope(Vehicle.all).purchased.pluck(:id)

          VehicleUpgrade.where(vehicle_id: vehicle_ids).delete_all
          VehicleModule.where(vehicle_id: vehicle_ids).delete_all
          Vehicle.where(id: vehicle_ids).delete_all
        end
      end

      def import
        authorize! to: :update?, with: ::HangarPolicy

        import = Imports::HangarImport.new(import_params.merge(user_id: current_user.id))

        unless import.save
          render json: ValidationError.new("hangar.import", errors: import.errors), status: :bad_request
          return
        end

        if import.import_data.blank?
          render json: {code: "hangar.import", message: I18n.t("errors.messages.hangar.import.no_data")}, status: :bad_request
          return
        end

        @response = ::HangarImporter.new(import).run
      rescue JSON::ParserError => e
        render json: ValidationError.new("hangar.import", message: e), status: :bad_request
      end

      def export
        authorize! to: :show?, with: ::HangarPolicy

        scope = authorized_scope(Vehicle.all).visible.purchased

        scope = loaner_included?(scope)

        vehicle_query_params["sorts"] = "model_name asc"

        @q = scope.ransack(vehicle_query_params)

        @vehicles = @q.result(distinct: true)
          .includes(model: [:manufacturer])
          .joins(model: [:manufacturer])
      end

      def sync_rsi_hangar
        authorize! to: :update?, with: ::HangarPolicy

        render json: ValidationError.new("vehicle.sync", message: I18n.t("messages.hangar_sync.no_data")), status: :bad_request if params[:items].blank?

        @response = ::HangarSync.new(sync_params[:items].to_a.map(&:to_h)).run(current_user.id)
      end

      def items
        authorize! to: :show?, with: ::HangarPolicy

        model_ids = authorized_scope(Vehicle.all)
          .where(loaner: false)
          .visible
          .purchased
          .pluck(:model_id)
        @models = Model.where(id: model_ids).order(name: :asc).pluck(:slug)
      end

      def move_all_ingame_to_wishlist
        authorize! to: :update?, with: ::HangarPolicy

        errors = []

        Vehicle.transaction do
          scope = authorized_scope(Vehicle.all).purchased.where(bought_via: :ingame)

          scope.find_each do |vehicle|
            next if vehicle.update(wanted: true)

            errors << vehicle.errors
          end
        end

        return if errors.blank?

        render json: ValidationError.new("vehicle.move_all_ingame_to_wish_list", errors:), status: :bad_request
      end

      private def import_params
        @import_params ||= params.permit(:import, :new_import)
      end

      private def sync_params
        @sync_params ||= params.permit(items: [:id, :name, :image, :type, :custom_name])
      end
    end
  end
end