# swagger/admin/v1/schema.yaml
...
openapi: 3.0.3
info:
  title: FleetYards.net Command API
  version: v1
  license:
    name: GNU General Public License v3.0
    url: https://github.com/fleetyards/fleetyards/blob/main/LICENSE
  x-logo:
    url: https://fleetyards.net/docs/logo.png
    altText: FleetYards.net logo
servers:
- url: http://admin.fleetyards.test/api/v1
  description: Dev Server
- url: http://admin.fleetyards.test/api/v1
  description: Production Server
- url: https://admin.fleetyards.dev/api/v1
  description: Staging Server
security:
- SessionCookie: []
paths:
...
  "/imports":
    get:
      summary: Imports list
      operationId: imports
      description: Get a List of Imports
      tags:
      - Imports
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - name: perPage
        in: query
        schema:
          type: string
          default: 25
        required: false
      - "$ref": "#/components/parameters/SortingParameter"
      - name: q
        in: query
        schema:
          type: object
          "$ref": "#/components/schemas/ImportQuery"
        style: deepObject
        explode: true
        required: false
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Imports"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/manufacturers":
    get:
      summary: Manufacturers list
      operationId: manufacturers
      tags:
      - Manufacturers
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - name: perPage
        in: query
        schema:
          type: string
          default: 30
        required: false
      - "$ref": "#/components/parameters/SortingParameter"
      - name: q
        in: query
        schema:
          type: object
          "$ref": "#/components/schemas/ManufacturerQuery"
        style: deepObject
        explode: true
        required: false
      - name: cacheId
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Manufacturers"
        '403':
          description: forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
  "/manufacturers/{id}":
    parameters:
    - name: id
      in: path
      schema:
        type: string
        format: uuid
      description: Manufacturer id
      required: true
    get:
      summary: Manufacturer Detail
      operationId: manufacturer
      tags:
      - Manufacturers
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Manufacturer"
        '404':
          description: not found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '403':
          description: forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
...
  "/vehicles":
    get:
      summary: Vehicles list
      operationId: vehicles
      tags:
      - Vehicles
      parameters:
      - "$ref": "#/components/parameters/PageParameter"
      - name: perPage
        in: query
        schema:
          type: string
          default: 30
        required: false
      - "$ref": "#/components/parameters/SortingParameter"
      - name: q
        in: query
        schema:
          type: object
          "$ref": "#/components/schemas/VehicleQuery"
        style: deepObject
        explode: true
        required: false
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Vehicles"
        '403':
          description: forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
        '401':
          description: unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/StandardError"
components:
  parameters:
    PageParameter:
      name: page
      in: query
      schema:
        type: string
        default: '1'
      required: false
    SortingParameter:
      name: s
      in: query
      schema:
        type: array
        items:
          type: string
      description: Sorting
      required: false
  schemas:
...
    BoughtViaEnum:
      type: string
      enum:
      - pledge_store
      - ingame
      title: BoughtViaEnum
...
    ImportStatusEnum:
      type: string
      enum:
      - created
      - started
      - finished
      - failed
      title: ImportStatusEnum
    ImportTypeEnum:
      type: string
      enum:
      - Imports::ModelImport
      - Imports::ModelsImport
      - Imports::ScData::AllImport
      - Imports::ScData::ModelsImport
      - Imports::ScData::ModelImport
      - Imports::HangarSync
      - Imports::HangarImport
      title: ImportTypeEnum
...
    MediaFile:
      type: object
      properties:
        name:
          type: string
        contentType:
          type: string
        size:
          type: integer
        url:
          type: string
          format: uri
        smallUrl:
          type: string
          format: uri
        mediumUrl:
          type: string
          format: uri
        largeUrl:
          type: string
          format: uri
        xlargeUrl:
          type: string
          format: uri
        width:
          type: integer
        height:
          type: integer
        uploadedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - name
      - contentType
      - size
      - url
      - uploadedAt
      title: MediaFile
    Meta:
      type: object
      properties:
        pagination:
          "$ref": "#/components/schemas/Pagination"
      additionalProperties: false
      title: Meta
...
    Pagination:
      type: object
      properties:
        totalCount:
          type: integer
        currentPage:
          type: integer
        totalPages:
          type: integer
        defaultPerPage:
          type: integer
        maxPerPage:
          type: integer
        perPageSteps:
          type: array
          items:
            anyOf:
            - type: string
            - type: integer
      additionalProperties: false
      required:
      - currentPage
      - totalPages
      title: Pagination
...
    VehicleSortEnum:
      type: string
      enum:
      - flagship desc
      - flagship asc
      - name asc
      - name desc
      - model_name asc
      - model_name desc
      - created_at asc
      - created_at desc
      - updated_at asc
      - updated_at desc
      - model_manufacturer_name asc
      - model_manufacturer_name desc
      - model_length asc
      - model_length desc
      - model_beam asc
      - model_beam desc
      - model_height asc
      - model_height desc
      - model_mass asc
      - model_mass desc
      - model_cargo asc
      - model_cargo desc
      - model_pledge_price asc
      - model_pledge_price desc
      - model_price asc
      - model_price desc
      - model_scm_speed asc
      - model_scm_speed desc
      - model_max_speed asc
      - model_max_speed desc
      - model_ground_max_speed asc
      - model_ground_max_speed desc
      - model_production_status asc
      - model_production_status desc
      - model_focus asc
      - model_focus desc
      title: VehicleSortEnum
    StandardError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      additionalProperties: false
      required:
      - code
      - message
      title: StandardError
...
    HangarGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        color:
          type: string
        public:
          type: boolean
        sort:
          type: integer
        vehiclesCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - name
      - slug
      - color
      - public
      - createdAt
      - updatedAt
      title: HangarGroup
...
    Import:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          "$ref": "#/components/schemas/ImportTypeEnum"
        status:
          "$ref": "#/components/schemas/ImportStatusEnum"
        info:
          type: string
        version:
          type: string
        input:
          type: string
        output:
          type: string
        import:
          type: string
        importData:
          type: string
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        failedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - type
      - status
      - createdAt
      - updatedAt
      title: Import
    Imports:
      type: object
      properties:
        meta:
          "$ref": "#/components/schemas/Meta"
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Import"
      additionalProperties: false
      required:
      - items
      - meta
      title: Imports
...
    Manufacturer:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        code:
          type: string
        logo:
          type: string
        longName:
          type: string
        scRef:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
      additionalProperties: false
      required:
      - name
      - slug
      - createdAt
      - updatedAt
      - id
      title: Manufacturer
    Manufacturers:
      type: object
      properties:
        meta:
          "$ref": "#/components/schemas/Meta"
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Manufacturer"
      additionalProperties: false
      required:
      - items
      - meta
      title: Manufacturers
...
    ModelUpgrade:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        pledgePrice:
          type: number
        media:
          type: object
          properties:
            storeImage:
              "$ref": "#/components/schemas/MediaFile"
          additionalProperties: false
        model:
          "$ref": "#/components/schemas/Model"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        storeImage:
          type: string
          format: uri
          deprecated: true
        storeImageLarge:
          type: string
          format: uri
          deprecated: true
        storeImageMedium:
          type: string
          format: uri
          deprecated: true
        storeImageSmall:
          type: string
          format: uri
          deprecated: true
      additionalProperties: false
      required:
      - id
      - name
      - media
      - model
      - createdAt
      - updatedAt
      title: ModelUpgrade
...
    ImportQuery:
      type: object
      properties:
        typeEq:
          "$ref": "#/components/schemas/ImportTypeEnum"
        typeIn:
          type: array
          items:
            "$ref": "#/components/schemas/ImportTypeEnum"
      additionalProperties: false
      title: ImportQuery
...
    ManufacturerQuery:
      type: object
      properties:
        withModels:
          type: boolean
        logoBlank:
          type: boolean
        nameEq:
          type: string
        nameCont:
          type: string
        nameIn:
          type: array
          items:
            type: string
        slugEq:
          type: string
        slugCont:
          type: string
        slugIn:
          type: array
          items:
            type: string
      additionalProperties: false
      title: ManufacturerQuery
...
    VehicleQuery:
      type: object
      properties:
        searchCont:
          type: string
        idEq:
          type: string
        idIn:
          type: array
          items:
            type: string
        idNotIn:
          type: array
          items:
            type: string
        nameCont:
          type: string
        nameIn:
          type: array
          items:
            type: string
        userUsernameIn:
          type: array
          items:
            type: string
        modelSlugIn:
          type: array
          items:
            type: string
        manufacturerIn:
          type: array
          items:
            type: string
        modelIdEq:
          type: string
        modelIdIn:
          type: array
          items:
            type: string
        modelIdNotIn:
          type: array
          items:
            type: string
        modelNameCont:
          type: string
        modelNameIn:
          type: array
          items:
            type: string
        modelProductionStatusIn:
          type: array
          items:
            type: string
        modelSearchCont:
          type: string
        loanerEq:
          type: boolean
        wantedEq:
          type: boolean
        sorts:
          anyOf:
          - type: array
            items:
              "$ref": "#/components/schemas/VehicleSortEnum"
          - "$ref": "#/components/schemas/VehicleSortEnum"
      additionalProperties: false
      title: VehicleQuery
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        unconfirmedEmail:
          type: string
        avatar:
          type: string
        rsiHandle:
          type: string
        discord:
          type: string
        youtube:
          type: string
        twitch:
          type: string
        guilded:
          type: string
        homepage:
          type: string
        saleNotify:
          type: boolean
        publicHangar:
          type: boolean
        publicHangarUrl:
          type: string
        publicHangarLoaners:
          type: boolean
        publicWishlist:
          type: boolean
        publicWishlistUrl:
          type: string
        hideOwner:
          type: boolean
        twoFactorRequired:
          type: boolean
        twoFactorQrCodeUrl:
          type: string
        twoFactorProvisioningUrl:
          type: string
        hangarUpdatedAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - username
      - email
      - saleNotify
      - publicHangar
      - publicHangarLoaners
      - publicWishlist
      - hideOwner
      - twoFactorRequired
      - createdAt
      - updatedAt
      title: User
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        serial:
          type: string
        user:
          "$ref": "#/components/schemas/User"
        alternativeNames:
          type: array
          items:
            type: string
        boughtVia:
          "$ref": "#/components/schemas/BoughtViaEnum"
        boughtViaLabel:
          type: string
        flagship:
          type: boolean
        hangarGroupIds:
          type: array
          items:
            type: string
            format: uuid
        hangarGroups:
          type: array
          items:
            "$ref": "#/components/schemas/HangarGroup"
        loaner:
          type: boolean
        model:
          "$ref": "#/components/schemas/Model"
        modelModuleIds:
          type: array
          items:
            type: string
            format: uuid
        modelUpgradeIds:
          type: array
          items:
            type: string
            format: uuid
        modulePackage:
          "$ref": "#/components/schemas/ModelModulePackage"
        nameVisible:
          type: boolean
        paint:
          "$ref": "#/components/schemas/ModelPaint"
        public:
          type: boolean
        saleNotify:
          type: boolean
        upgrade:
          "$ref": "#/components/schemas/ModelUpgrade"
        wanted:
          type: boolean
        hidden:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
      required:
      - id
      - model
      - user
      - wanted
      - hidden
      - boughtVia
      - loaner
      - flagship
      - public
      - nameVisible
      - saleNotify
      - alternativeNames
      - modelUpgradeIds
      - hangarGroupIds
      - hangarGroups
      - modelModuleIds
      - createdAt
      - updatedAt
      title: Vehicle
    Vehicles:
      type: object
      properties:
        meta:
          "$ref": "#/components/schemas/Meta"
        items:
          type: array
          items:
            "$ref": "#/components/schemas/Vehicle"
      additionalProperties: false
      required:
      - items
      - meta
      title: Vehicles
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    Oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://fleetyards.net/oauth/authorize
          tokenUrl: https://fleetyards.net/oauth/token
          scopes:
            public: Public Access
            profile:read: Read Profile
            profile:write: Manage Profile
            hangar: Full Hangar Access
            hangar:read: Read Hangar
            hangar:write: Manage Hangar
            fleet: Full Fleet Access
            fleet:read: Read Fleet
            fleet:write: Manage Fleet
            openid: Authenticate your account
    OpenId:
      type: openIdConnect
      openIdConnectUrl: http://fleetyards.test/.well-known/openid-configuration
    SessionCookie:
      type: apiKey
      description: Session Cookie
      name: FLTYRD
      in: cookie